requirements.txt
aenum==3.1.15
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
dateparser==1.2.0
fonttools==4.53.0
html5lib==1.1
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
simplejson==3.19.2
six==1.16.0
tinycss2==1.3.0
tzlocal==5.2
urllib3==2.2.1
wallbox==0.6.0
weasyprint==62.1
webencodings==0.5.1
zopfli==0.2.3
